+++ CHANGELOG.md
@@ -2349,3 +2349,4 @@
 - [2026-06-26 09:28] feat(state): add async state stream provider
 - [2026-06-26 09:33] feat(auth): add async auth stream provider
 - [2026-06-26 09:43] test(auth): add unit tests for auth business logic
+- [2026-06-26 09:45] refactor(cache): optimize data flow between domain and data layers